Add simpler logic for languages using same open and close quote
This works well for the case of Finnish and Swedish where we use:
("”" "”" "’" "’")

Now, invoking typo-insert-quotation-mark with " cycles between
opening-double-quote and typewriter quote, and for using single
quotes, Swedish and Finnish writers can use ' to invoke
typo-cycle-right-single-quotation-mark (which will be the right symbol
for these languages).

Additionally add an entry for Swedish.